requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
pywin32
pillow
requests
pydantic
python-multipart
transformers
numpy
scikit-image